third_party/WebKit/LayoutTests/fast/mediastream/MediaStream-add-remove-tracks-expected.txt

Test adding and removing tracks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS stream1.getAudioTracks().length is 1
PASS stream1.getVideoTracks().length is 1
PASS stream1.getTracks().length is 2
PASS stream2.getAudioTracks().length is 1
PASS stream2.getVideoTracks().length is 1
PASS stream2.getTracks().length is 2
PASS stream1.getAudioTracks().length is 0
PASS stream1.getTracks().length is 1
PASS stream1.getTrackById(audioTrack.id) is null
PASS stream1.getTrackById(audioTrack.id) is non-null.
PASS audioTrack is stream1.getAudioTracks()[0]
PASS stream1.getVideoTracks().length is 0
PASS stream1.getTracks().length is 1
PASS stream1.getTrackById(videoTrack.id) is null
PASS stream1.getAudioTracks().length is 1
PASS stream1.getVideoTracks().length is 1
PASS stream1.getTracks().length is 2
PASS stream1.getAudioTracks()[0].kind is audioTrack.kind
PASS stream1.getAudioTracks()[0].label is audioTrack.label
PASS stream1.getVideoTracks()[0].kind is videoTrack.kind
PASS stream1.getVideoTracks()[0].label is videoTrack.label
PASS stream1.getAudioTracks().length is 1
PASS stream1.getVideoTracks().length is 1
PASS stream1.getTracks().length is 2
PASS stream1.getAudioTracks().length is 2
PASS stream1.getVideoTracks().length is 2
PASS stream1.getTracks().length is 4
PASS stream2.ended is true
PASS stream1.addTrack(audioTrack) threw exception InvalidStateError: Failed to execute 'addTrack' on 'MediaStream': The MediaStream is finished..
PASS stream1.removeTrack(audioTrack) threw exception InvalidStateError: Failed to execute 'removeTrack' on 'MediaStream': The MediaStream is finished..
PASS successfullyParsed is true

TEST COMPLETE
